split tests and add test README
