+++ test/README.md
@@ -1,0 +1,15 @@
+# Tests
+
+Test are written in JUnit and Maven
+
+Run with Maven: 
+
+```bash
+mvn install
+```
+
+Run with Docker-Compose
+
+```bash
+docker-compose build && docker-compose run test
+```
